move fairness to table subpackage + flagrate instead of flagged
