+++ src/seismometer/table/fairness.py
